requirements.txt
Pillow>=9.0.0
platformdirs
pyyaml
panda3d-blend2bam
numpy~=1.24.0
